refactor(agent-runtime): extract provider contracts to corvus-traits
