+++ clients/agent-runtime/crates/corvus-traits/Cargo.toml
@@ -10,8 +10,11 @@ repository = "https://github.com/dallay/corvus"
 [dependencies]
 anyhow = "1.0"
 async-trait = "0.1"
+futures-util = { version = "0.3", default-features = false, features = ["alloc"] }
+reqwest = { version = "0.12", default-features = false }
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 serde_json = { version = "1.0", default-features = false, features = ["std"] }
+thiserror = "2.0"
 tokio = { version = "1.42", default-features = false, features = ["sync"] }
 tracing = { version = "0.1", default-features = false }
