+++ clients/agent-runtime/crates/corvus-traits/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod channels;
 pub mod memory;
 pub mod multimedia;
+pub mod providers;
 pub mod security;
 mod testing;
 pub mod tools;
